requirements.txt
# requirements.txt
openai
requests
tenacity
google-genai 
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index
pinecone-client
llama-index-vector-stores-pinecone